Second checkpoint

+++ DB.py
@@ -1,6 +1,6 @@
 import sys
 import time
-from typing import re
+import re
 
 import mysql.connector
 import pandas as pd
@@ -13,7 +13,7 @@ my_db = mysql.connector.connect(
     database='Scraper'
 )
 
-my_cursor = my_db.cursor()
+my_cursor = my_db.cursor(buffered=True)
 
 
 def create_db():
@@ -73,21 +73,39 @@ def first_fill(data: pd.core.frame.DataFrame):
 
 def update_city(city, df_new):
     # Delete apartments that are no longer in the page
-    df_old = get_query_df('SELECT * FROM place WHERE city = "https://www.waytostay.com/' + city + '-apartments/"')
+    try:
+        df_old = get_query_df('SELECT * FROM place WHERE city = "https://www.waytostay.com/' + city + '-apartments/"')
+    except mysql.connector.errors.InternalError:
+        df_old = pd.DataFrame()
+        print('Inserting new city')
+
     new_apartments = (str(list(df_new['page_link'].values))).strip('[]')
-    query_1 = "DELETE FROM place WHERE page_link NOT IN ({0})".format(new_apartments)
-    my_cursor.execute(query_1)
+
+    if not df_old.empty:
+        query_1 = "DELETE FROM place WHERE page_link NOT IN ({0})".format(new_apartments)
+        my_cursor.execute(query_1)
 
     # Insert new apartments
     num_col = df_new.columns
-    df_insert = df_new.merge(df_old, on='page_link', how='left', indicator=True).query('_merge == "left_only"').drop('_merge', 1)
-    df_insert = df_insert[df_insert.columns[0:len(df_new.columns)]]
-    df_insert.columns = num_col
+    if not df_old.empty:
+        df_insert = df_new.merge(df_old, on='page_link', how='left', indicator=True).query('_merge == "left_only"').drop('_merge', 1)
+        df_insert = df_insert[df_insert.columns[0:len(df_new.columns)]]
+        df_insert.columns = num_col
+    else:
+        df_insert = df_new
     data, data_prep_list, curr_list = prep_data(df_insert)
-    my_cursor.execute("SELECT MAX(home_id) FROM place")
-    max_id = my_cursor.fetchall()
-    for i in range(max_id[0][0] + 1, max_id[0][0] + 1 + len(data_prep_list)):
-        data_prep_list[i-max_id[0][0] - 1].insert(0, i)
+    try:
+        my_cursor.execute("SELECT MAX(home_id) FROM place")
+        max_id = my_cursor.fetchall()
+        for i in range(max_id[0][0] + 1, max_id[0][0] + 1 + len(data_prep_list)):
+            data_prep_list[i-max_id[0][0] - 1].insert(0, i)
+    except TypeError:
+        for i in range(len(data_prep_list)):
+            data_prep_list[i].insert(0, i)
+
+    except mysql.connector.errors.InternalError:
+        for i in range(len(data_prep_list)):
+            data_prep_list[i].insert(0, i)
     query_2 = """INSERT INTO place (
                                 home_id,
                                 city ,
@@ -102,43 +120,48 @@ def update_city(city, df_new):
 
     # Update apartments whose data has been changed
 
-    df_update = df_new.merge(df_old, on='page_link', how='inner', indicator=True)
-    df_update = df_update[df_update.columns[0:len(df_new.columns)]]
-    df_update.columns = num_col
+    if not df_old.empty:
+        df_update = df_new.merge(df_old, on='page_link', how='inner', indicator=True)
+        df_update = df_update[df_update.columns[0:len(df_new.columns)]]
+        df_update.columns = num_col
 
-    upd_apartments = (str(list(df_update['page_link'].values))).strip('[]')
-    query_3 = "DELETE FROM place WHERE page_link NOT IN ({0})".format(upd_apartments)
-    my_cursor.execute(query_3)
+        upd_apartments = (str(list(df_update['page_link'].values))).strip('[]')
+        query_3 = "DELETE FROM place WHERE page_link NOT IN ({0})".format(upd_apartments)
+        my_cursor.execute(query_3)
 
-    data, data_prep_list, curr_list = prep_data(df_update)
-    my_cursor.execute("SELECT MAX(home_id) FROM place")
-    max_id = my_cursor.fetchall()
-    for i in range(max_id[0][0] + 1, max_id[0][0] + 1 + len(data_prep_list)):
-        data_prep_list[i - max_id[0][0] - 1].insert(0, i)
-    query_4 = """INSERT INTO place (
-                                    home_id,
-                                    city ,
-                                    page_link ,
-                                    sleeps ,
-                                    area_sqm ,
-                                    bedrooms , 
-                                    bathroom ,
-                                    price ,
-                                    curency_ID ) VALUES (%s, %s,%s,%s,%s, %s, %s, %s, %s)"""
-    my_cursor.executemany(query_4, data_prep_list)
+        data, data_prep_list, curr_list = prep_data(df_update)
+        my_cursor.execute("SELECT MAX(home_id) FROM place")
+        max_id = my_cursor.fetchall()
+        for i in range(max_id[0][0] + 1, max_id[0][0] + 1 + len(data_prep_list)):
+            data_prep_list[i - max_id[0][0] - 1].insert(0, i)
+        query_4 = """INSERT INTO place (
+                                        home_id,
+                                        city ,
+                                        page_link ,
+                                        sleeps ,
+                                        area_sqm ,
+                                        bedrooms , 
+                                        bathroom ,
+                                        price ,
+                                        curency_ID ) VALUES (%s, %s,%s,%s,%s, %s, %s, %s, %s)"""
+        my_cursor.executemany(query_4, data_prep_list)
     my_db.commit()
 
 
 def update_global(df_new):
     try:
+        my_cursor.execute("SELECT MAX(home_id) FROM place")
+    except mysql.connector.errors.InternalError:
         create_db()
         create_tables()
     finally:
+        i=1
         for city in df_new['city'].unique():
             city_name = re.search(r'https://www.waytostay.com/([a-z]*)-apartments/', str(city)).group(1)
             print(city_name)
             update_city(city_name, df_new[df_new['city'] == city])
-
+            if i == 1:
+                break
 
 def get_query_df(query):
     df = pd.read_sql_query(query, my_db)
